requirements.txt
absl-py==1.3.0
addict==2.4.0
asttokens==2.2.0
attrs==22.1.0
backcall==0.2.0
bm3d==4.0.1
bm4d==4.2.3
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
cloudpickle==2.2.0
codecov==2.1.12
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
coverage==6.5.0
cycler==0.11.0
debugpy==1.6.4
decorator==4.4.2
dm-tree==0.1.8
entrypoints==0.4
etils==0.9.0
exceptiongroup==1.0.1
executing==1.2.0
filelock==3.8.0
flax==0.6.3
flow-vis==0.1
fonttools==4.38.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.50.0
huggingface-hub==0.11.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.8
importlib-metadata==5.0.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.7.0
jax==0.3.17
jaxlib==0.3.15
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.1.0
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mmcv==1.7.0
mmdet==2.25.3
model-index==0.1.11
moviepy==1.0.3
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.6.0.66
openmim==0.3.3
opt-einsum==3.3.0
optax==0.1.4
orbax==0.0.23
ordered-set==4.1.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
proglog==0.1.10
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAbel==0.9.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scico==0.0.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
stack-data==0.6.2
submitit==1.4.5
svmbir==0.3.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorstore==0.1.28
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.3.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
torchviz==0.0.2
tornado==6.2
tqdm==4.65.0
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.10.0
